3 Parameter tracking: series of alterations to fix memory leak.
For 3 parameter fitting minimisation is now performed by
ROOT::Minuit2::Minuit2Minimizer that is set-up by a pointer in the
THaVDCCluster class header. The functor used is similarly set-up as a
pointer in the THaVDCCluster class header.

To use ROOT::Minuit2::Minuit2Minimizer the lMinuit2 library had to be
added to the makefile.