Added option to cut tracks with 'ambiguous' matches between chambers
This is controlled by parameter read in via VDC DB. Parameter is
AmbiClusters (a bool) which if true cuts tracks where one of the
cluster pairs (U and V cluster in one chamber) points to more than one
clusrter pair in the other chamber. Points to here means that it is
within the cutoff limits.