VDC: Added test if clusters are within APEX X,Y range.
Function added (to THaVDCPlane) to test if given x and y (in DCS) are
within range of APEX distributions. Stricter version of existing
IsInActiveArea function.

Added CalcDetCoordsVPlane to THaVDCChamber, equivalent to existing
CalcDetCoords but takes U and V clusters and projects to V
plane. This allows additional test (IsInUsedArea or IsInActiveArea)
for pair.